pyproject.toml
# pyproject.toml
# This file is used exclusively for python configurations, including:
# * Dependency management
# * Static analysis configuration (formatting, linting, type checking)
# * High-level testing configuration
# * Code packaging and build configurations

[project]
name = "app"
version = "0.0.0"

dependencies = [
    "fastapi[standard]==0.115.4",
    "SQLModel==0.0.22",
    "SQLAlchemy==2.0.36",
    "pydantic==2.9.2",
    "pydantic-settings==2.6.1",
    "psycopg==3.2.3",
]

[project.optional-dependencies]
dev = [
    "python-dotenv==1.0.1",
    "typer==0.13.0",
    "black==24.10.0",
    "ruff==0.7.3",
    "mypy==1.11.2",
    "pytest==8.3.3",
    "types==PyYAML-6.0.12.20240917",
]

# Additional build options
[tool.setuptools]
py-modules = []

# Formatting python code through tool: black
# Configurations for static code formatter "black" -- using default configurations
[tool.black]
line-length = 88

# Static analysis through tool ruff
[tool.ruff.lint]
select = [
    "E", # pycodestyle
    "F", # pyflakes
    "UP", # pyupgrade
    "B", # flake8-bugbear
    "SIM", # flake8-simplify
    "I", # isort
]

# Type checking through tool: mypy

[tool.mypy]
plugins = "pydantic.mypy"

ignore_missing_imports = false
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
no_implicit_optional = true
strict_optional = true
strict_equality = true
check_untyped_defs = true
disallow_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true